+++ custom_cartridge_master/cartridges/custom_sanika1/cartridge/controllers/Cart.js
@@ -3,13 +3,8 @@
 /**
  * @namespace Cart
  */
-
 var server = require('server');
-var csrfProtection = require('*/cartridge/scripts/middleware/csrf');
-var consentTracking = require('*/cartridge/scripts/middleware/consentTracking');
-
 server.extend(module.superModule);
-
 server.replace('AddProduct', function (req, res, next) {
     var BasketMgr = require('dw/order/BasketMgr');
     var Resource = require('dw/web/Resource');
@@ -19,6 +14,7 @@ server.replace('AddProduct', function (req, res, next) {
     var ProductLineItemsModel = require('*/cartridge/models/productLineItems');
     var cartHelper = require('*/cartridge/scripts/cart/cartHelpers');
     var basketCalculationHelpers = require('*/cartridge/scripts/helpers/basketCalculationHelpers');
+    
     var currentBasket = BasketMgr.getCurrentOrNewBasket();
     var previousBonusDiscountLineItems = currentBasket.getBonusDiscountLineItems();
     var productId = req.form.pid;
@@ -27,7 +23,6 @@ server.replace('AddProduct', function (req, res, next) {
     var firstName=req.form.firstName;
     var lastName=req.form.lastName;
     var email=req.form.email;
-
     var childProducts = Object.hasOwnProperty.call(req.form, 'childProducts')
         ? JSON.parse(req.form.childProducts)
         : [];
@@ -87,13 +82,11 @@ server.replace('AddProduct', function (req, res, next) {
     }
     var quantityTotal = ProductLineItemsModel.getTotalQuantity(currentBasket.productLineItems);
     var cartModel = new CartModel(currentBasket);
-
     var urlObject = {
         url: URLUtils.url('Cart-ChooseBonusProducts').toString(),
         configureProductstUrl: URLUtils.url('Product-ShowBonusProducts').toString(),
         addToCartUrl: URLUtils.url('Cart-AddBonusProducts').toString()
     };
-
     var newBonusDiscountLineItem = cartHelper.getNewBonusDiscountLineItem(
         currentBasket,
         previousBonusDiscountLineItems,
@@ -112,9 +105,7 @@ server.replace('AddProduct', function (req, res, next) {
             }
         });
     }
-
     var reportingURL = cartHelper.getReportingUrlAddToCart(currentBasket, result.error);
-
     res.json({
         reportingURL: reportingURL,
         quantityTotal: quantityTotal,
@@ -137,10 +128,8 @@ server.replace(
         var cartHelper = require('*/cartridge/scripts/cart/cartHelpers');
         var reportingUrlsHelper = require('*/cartridge/scripts/reportingUrls');
         var basketCalculationHelpers = require('*/cartridge/scripts/helpers/basketCalculationHelpers');
-
         var currentBasket = BasketMgr.getCurrentBasket();
         var reportingURLs;
-
         if (currentBasket) {
             Transaction.wrap(function () {
                 if (currentBasket.currencyCode !== req.session.currency.currencyCode) {
@@ -151,17 +140,13 @@ server.replace(
                 basketCalculationHelpers.calculateTotals(currentBasket);
             });
         }
-
         if (currentBasket && currentBasket.allLineItems.length) {
             reportingURLs = reportingUrlsHelper.getBasketOpenReportingURLs(currentBasket);
         }
-
         res.setViewData({ reportingURLs: reportingURLs });
-
         var basketModel = new CartModel(currentBasket);
         res.render('cart/cart', basketModel);
         next();
     }
 );
-
 module.exports = server.exports();
